Connected to FDA API

+++ src/main/java/com/project/inventory_management/service/UserService.java
@@ -7,17 +7,21 @@ import com.project.inventory_management.dto.UserResponseDTO;
 import com.project.inventory_management.entity.User;
 import com.project.inventory_management.exception.UserExistException;
 import com.project.inventory_management.exception.UserNotFoundException;
+import com.project.inventory_management.exception.UserUnauthorizedException;
 import com.project.inventory_management.mapper.UserMapper;
 import com.project.inventory_management.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.sql.Timestamp;
 import java.time.Instant;
+import java.util.Collection;
 import java.util.Optional;
 
 @Service
@@ -80,13 +84,11 @@ public class UserService {
 
     }
 
-    @PreAuthorize("hasRole('ADMIN')")
     public UserResponseDTO getUserById(int id) {
         return userMapper.toUserResponseDTO(userRepository.findById(id)
                 .orElseThrow(() -> new UserNotFoundException("User not found with id: " + id)));
     }
 
-    @PreAuthorize("hasRole('ADMIN')")
     public UserResponseDTO updateUser(
             UserRequestDTO.UpdateUserRequestDTO updateUserRequestDTO,
             Integer id) {
@@ -109,10 +111,16 @@ public class UserService {
         throw new UserNotFoundException("User not found");
     }
 
-    @PreAuthorize("hasRole('ADMIN')")
     public Boolean deleteUserById(int id) {
         User userToDelete = userRepository.findById(id).orElseThrow(
                 () -> new UserNotFoundException("User not found with id: " + id));
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
+        boolean isAdmin = authorities.stream()
+                        .anyMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("ROLE_ADMIN"));
+        if (!isAdmin) {
+            throw new UserUnauthorizedException("User is not admin");
+        }
         userRepository.delete(userToDelete);
         return true;
     }
